requirements.txt
Django==2.2.2
matplotlib==3.1.0
mpld3==0.3
Jinja2==2.10.1
netCDF4==1.5.1.2
pptk==0.1.0
#useless in future
ivy-python==3.1